Move pairs to external file, join address file in pairs.js
